Trekker ut logikk for opprette oppgave hvis alderspensjon i egen service og justerer logikk
hvis det er førstegangsbehandling skal den sjekke om det er overstyrt
grunnet alderspensjon inneværende og neste år
