+++ apps/etterlatte-beregning/src/main/kotlin/config/ApplicationContext.kt
@@ -4,6 +4,7 @@ import com.typesafe.config.Config
 import com.typesafe.config.ConfigFactory
 import no.nav.etterlatte.avkorting.AvkortingRepository
 import no.nav.etterlatte.avkorting.AvkortingService
+import no.nav.etterlatte.avkorting.AvkortingTidligAlderspensjonService
 import no.nav.etterlatte.beregning.AnvendtTrygdetidRepository
 import no.nav.etterlatte.beregning.BeregnBarnepensjonService
 import no.nav.etterlatte.beregning.BeregnOmstillingsstoenadService
@@ -111,6 +112,11 @@ class ApplicationContext {
             sanksjonService = sanksjonService,
             grunnlagKlient = grunnlagKlient,
             featureToggleService = featureToggleService,
+        )
+    val avkortingTidligAlderspensjonService =
+        AvkortingTidligAlderspensjonService(
+            behandlingKlient = behandlingKlient,
+            avkortingRepository = avkortingRepository,
         )
     val ytelseMedGrunnlagService =
         YtelseMedGrunnlagService(
