+++ apps/etterlatte-beregning/src/test/kotlin/avkorting/AvkortingRoutesTest.kt
@@ -42,6 +42,7 @@ class AvkortingRoutesTest {
     private val mockOAuth2Server = MockOAuth2Server()
     private val behandlingKlient = mockk<BehandlingKlient>()
     private val avkortingService = mockk<AvkortingService>()
+    private val avkortingTidligAlderspensjonService = mockk<AvkortingTidligAlderspensjonService>()
 
     @BeforeAll
     fun beforeAll() {
@@ -179,7 +180,7 @@ class AvkortingRoutesTest {
     private fun testApplication(block: suspend ApplicationTestBuilder.() -> Unit) {
         io.ktor.server.testing.testApplication {
             runServer(mockOAuth2Server) {
-                avkorting(avkortingService, behandlingKlient)
+                avkorting(avkortingService, behandlingKlient, avkortingTidligAlderspensjonService)
             }
             block(this)
         }
